theme, ui: move styling data into `ui`
